TINY-9977: Extracted Schema sets and lookups from API module
